LayoutTests/platform/chromium/fast/forms/suggestion-picker/week-suggestion-picker-reset-value-after-reload-expected.txt

Checks reloading page resets input values

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS highlightedEntry() is null
PASS highlightedEntry() is "2012-W01"
PASS document.getElementById("mock-page-popup") is null
PASS iframe.contentDocument.getElementById("test1").value is "2012-W01"
PASS highlightedEntry() is null
PASS highlightedEntry() is "2012-W01"
PASS document.getElementById("mock-page-popup") is null
PASS iframe.contentDocument.getElementById("test2").value is "2012-W01"
PASS testInput1.value is ""
PASS testInput2.value is "2002-W02"
PASS successfullyParsed is true

TEST COMPLETE
